// SpringBoot/BlogBackend/src/app/shared/model/course.ts
import { CourseLocation } from './courseLocation';
import { CourseApply } from './courseApply';
import { Files } from './files';
export class Course {
  courseId: number;
  courseType: number;
  courseName: string;
  courseDescription: null;
  courseHours: number;
  applyStartDate: Date;
  applyEndDate: Date;
  teacherName: string;
  organizerUnit: string;
  connectionUnit: string;
  connectionName: string;
  connectionTel: string;
  connectionEmail: string;
  checkinRules: number;
  isMeals: number;
  mealsRemarks: string;
  isApply: number;
  isFillout: number;
  courseQrcode: string;
  postStartDate: Date;
  postEndDate: Date;
  courseRemarks: string;
  userCode: string;
  applicantlimit: number;
  feedbackId: number;
  certificateId: number;
  certificateYear: number;
  className: string;
  creationDate: Date;
  createdBy: string;
  updateDate: Date;
  updatedBy: string;
  courseApplyList: CourseApply[];
  courseLocationList: CourseLocation[];
  filesList: Files[];
}